Add files via upload
This website offers a easy and complete service to track the public transportation routes including the timeline feature that shows the real-time movement of the bus. Users can easily pick their start location, destination stop, and the bus of their choice to get the most correct travel details.

To make the platform more reachable, it is a trilingual platform that supports Hindi, English, and Punjabi.

In order to ensure the safety of users, emergency call buttons are installed in the device for instant help in case of need. Moreover, a live chatbox is always there to provide assistance and answer the doubts of users.

The website from which users can also access light and dark mode caters to user’s personalized and eye-friendly surfing experience.